COMP2411/Project/OSS Project/test/instructionTest/AdministratorTest/AdministratorGetAllOrderExecutorTest.java
package instructionTest.AdministratorTest;

import model.instruction.administrator.AdministratorGetAllOrderExecutor;
import org.junit.Before;
import org.junit.Test;

import java.io.ByteArrayOutputStream;
import java.io.PrintStream;
import java.sql.SQLException;

import static org.junit.Assert.assertEquals;

public class AdministratorGetAllOrderExecutorTest {

    private AdministratorGetAllOrderExecutor executor;
    private final ByteArrayOutputStream outContent = new ByteArrayOutputStream();


    @Before
    public void setUpStreams() {
        System.setOut(new PrintStream(outContent));
    }

    @Before
    public void setUp() {
        executor = new AdministratorGetAllOrderExecutor("1234567890");
    }

    @Test
    public void testInformation() throws SQLException {
        executor.information();
        String expectedOutput = "*********************************************************************************\n" +
                "You are now viewing all of the orders added in the OSS system\n\n";
        assertEquals(expectedOutput, outContent.toString());
    }


    @Test
    public void testChoose() throws SQLException {
        executor.choose("input");
    }

    @Test
    public void testVisitDatabase() throws SQLException {
        executor.visitDatabase();
    }
}

